Sitemap finished
Sitemap is done, but it could probably use some polish at a later date
such as minimizing the height of the main image on the page, in fact,
all of the heights on these images should follow a specific aspect ratio
so as not to throw off the user. also did some work on the staff and
board pages. Still need updated content for these pages, but that will
have to wait until after the December meeting.