First steps login
First steps at implementing a login function. Working so far, but still
needs polishing. The idea is to implement one login feature for the
whole site and based on permissions show or don't show certain aspects
of the site. For example, for staff members in the staff group, there
will be a link to fill out daily/weekly accountability entries whereas
for any user without the group, there will not.